+++ src/compas_timber/ghpython/components/CT_AutomaticJoints/code.py
@@ -2,6 +2,7 @@ from ghpythonlib.componentbase import executingcomponent as component
 from Grasshopper.Kernel.GH_RuntimeMessageLevel import Warning
 
 from compas_timber.connections import ConnectionSolver
+from compas_timber.connections import JointTopology
 from compas_timber.ghpython import JointDefinition
 
 
@@ -19,19 +20,25 @@ class AutotomaticJoints(component):
         rules = [r for r in Rules if r is not None]
 
         solver = ConnectionSolver()
-        found_pairs = solver.find_intersecting_pairs(Beams, rtree=True)
+        found_pairs = solver.find_intersecting_pairs(Beams, rtree=True, max_distance=MaxDistance)
         JointDefs = []
         Info = []
         # rules have to be resolved into joint definitions
         for pair in found_pairs:
             beam_a, beam_b = pair
             detected_topo, beam_a, beam_b = solver.find_topology(beam_a, beam_b, max_distance=MaxDistance)
+
+            if detected_topo == JointTopology.TOPO_UNKNOWN:
+                continue
+
             for rule in rules:
                 if not rule.comply(pair):
                     continue
                 if rule.joint_type.SUPPORTED_TOPOLOGY != detected_topo:
                     msg = "Conflict detected! Beams: {}, {} meet with topology: {} but rule assigns: {}"
-                    Info.append(msg.format(beam_a.key, beam_b.key, detected_topo, rule.joint_type.__name__))
+                    Info.append(
+                        msg.format(beam_a, beam_b, JointTopology.get_name(detected_topo), rule.joint_type.__name__)
+                    )
                     continue
                 # sort by category to allow beam role by order (main beam first, cross beam second)
                 beam_a, beam_b = rule.reorder([beam_a, beam_b])